reorganized image files/folders
