+++ client/src/assets/images/login-signup/labels/login-animated.gif
