+++ client/src/assets/images/login-signup/labels/signup.png
