+++ client/src/pages/LandingPage.jsx
@@ -10,7 +10,7 @@ const LandingPage = () => {
   // import into App to show on load '/'
   // Giant Fight My Dad img in the middle of page
   return (
-    <div>
+    <div style={{ backgroundImage: {} }}>
       <Container>
         <Grid>
           <Button
